Components/NavigationMenuViewComponent.cs
using System;
using System.Collections.Generic;
using System.Linq;
using Microsoft.AspNetCore.Mvc;
using SportsStore.Models;

namespace SportsStore.Components
{
    public class NavigationMenuViewComponent : ViewComponent
    {
        private IProductRepository _repository;

        public NavigationMenuViewComponent(IProductRepository repository)
        {
            _repository = repository;
        }

        public IViewComponentResult Invoke()
        {
            ViewBag.SelectedCategory = RouteData?.Values["category"]?.ToString();

            var productGroups = (from product in _repository.Products
                group product by product.Category).ToDictionary(g => g.Key, g => g.Count());

            return View(productGroups);
        }
    }
}